squibble: be smarter about diffing SQL-only objects
SQLite's "normalization" is not that robust, and in particular does not really
handle line breaks in a consistent way. Semantically this is fine, but it can
result in a spurious diff for a view (say), since

   CREATE VIEW foo AS SELECT x, y FROM bar;

and

   CREATE VIEW foo AS
     SELECT x, y
       FROM bar;

will be "different" after normalization, but produce the same view.  So before
applying the text diff, check the "collapsed" cleaned SQL with newlines folded
out first, and only print a diff if they differ there.