squibble: allow filtering of unmanaged tables (#2)
In some cases, a database may have tables that squibble should not treat as
part of the update schema. We already have a couple examples of this, we filter
out _schema_history (which squibble populates), and the sqlite_sequence table.

Another example, as described in #1, are the tables created by the Litestream
replication tool (https://litestream.io/).

Add a field to the Schema type allowing the user to specify database and view
names that should be excluded, and include those in the filter we apply when
loading a schema for comparison.